refactor: separate generateAutomaticTickets and generateManualTickets to a separate class for better testability
